refactor: rename the MilestoneTask folder
