+++ wondrous-app/components/Common/MilestoneTask/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './MilestoneTasks';
